Mother load of changes
Add gif and APNG recording(APNG untested and GIF broken in opengl)
Screenshots options menu
Backported the zone memory mangament system from 2.1
Uhh bunch of other changes too

+++ src/i_system.h
@@ -48,7 +48,7 @@ extern byte keyboard_started;
 
 	
 */
-ULONG I_GetFreeMem(ULONG *total);
+UINT32 I_GetFreeMem(UINT32 *total);
 
 /**	\brief  Called by D_SRB2Loop, returns current time in tics.
 */
